pyproject.toml
[tool.poetry]
name = "python-meme-generator"
version = "0.1.0"
description = ""
authors = ["apkawa <apkawa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pycairo = "^1.19.1"
emoji = "^0.5.4"
PyGObject = "^3.36.0"
webcolors = "^1.11.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
pytest-mypy = "^0.6.0"
pytest-image-diff = {git = "https://github.com/Apkawa/pytest-image-diff"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
